+++ tests/test_blackList.py
@@ -15,7 +15,7 @@ def test_run_blackList_big():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': None, 'afthr': None, 'bigfile': 'tests/files/input_blackList.big', 'verbose': None}
+            'aftag': None, 'afthr': None, 'bigfile': 'tests/files/input_blackList.big', 'BEDfile': None, 'verbose': None}
     # Run
     main_blackList(args)
     # Tests
@@ -28,7 +28,7 @@ def test_run_blackList_032_big():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': 'novoAF', 'afthr': '0.32', 'bigfile': 'tests/files/input_blackList.big', 'verbose': None}
+            'aftag': 'novoAF', 'afthr': '0.32', 'bigfile': 'tests/files/input_blackList.big', 'BEDfile': None, 'verbose': None}
     # Run
     main_blackList(args)
     # Tests
@@ -41,7 +41,7 @@ def test_run_blackList_032():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': 'novoAF', 'afthr': '0.32', 'bigfile': None, 'verbose': None}
+            'aftag': 'novoAF', 'afthr': '0.32', 'bigfile': None, 'BEDfile': None, 'verbose': None}
     # Run
     main_blackList(args)
     # Tests
@@ -54,7 +54,7 @@ def test_run_blackList_01():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': 'novoAF', 'afthr': '0.1', 'bigfile': None, 'verbose': None}
+            'aftag': 'novoAF', 'afthr': '0.1', 'bigfile': None, 'BEDfile': None, 'verbose': None}
     # Run
     main_blackList(args)
     # Tests
@@ -67,11 +67,24 @@ def test_run_blackList_001():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': 'novoAF', 'afthr': '0.01', 'bigfile': None, 'verbose': None}
+            'aftag': 'novoAF', 'afthr': '0.01', 'bigfile': None, 'BEDfile': None, 'verbose': None}
     # Run
     main_blackList(args)
     # Tests
     assert [row for row in open('tests/files/main_test.out')] == [row for row in open('tests/files/input_blackList_0-01.out')]
+    # Clean
+    os.remove('tests/files/main_test.out')
+#end def
+
+def test_run_blackList_BED():
+    ''' '''
+    # Variables
+    args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
+            'aftag': None, 'afthr': None, 'bigfile': None, 'BEDfile': 'tests/files/input_BED_blacklist.bed', 'verbose': None}
+    # Run
+    main_blackList(args)
+    # Tests
+    assert [row for row in open('tests/files/main_test.out')] == [row for row in open('tests/files/input_blackList_BED.out')]
     # Clean
     os.remove('tests/files/main_test.out')
 #end def
@@ -84,18 +97,18 @@ def test_args_afthr_bigfile_missing():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': None, 'afthr': None, 'bigfile': None, 'verbose': None}
+            'aftag': None, 'afthr': None, 'bigfile': None, 'BEDfile': None, 'verbose': None}
     # Run and Tests
     with pytest.raises(SystemExit) as e:
         assert main_blackList(args)
-    assert str(e.value) == '\nERROR in parsing arguments: to blacklist specify a BIG file and/or a threshold for population allele frequency and the TAG to use\n'
+    assert str(e.value) == '\nERROR in parsing arguments: to blacklist specify at least a BIG file, a BED file, and/or a threshold for population allele frequency and the TAG to use\n'
 #end def
 
 def test_args_afthr_conflict():
     ''' '''
     # Variables
     args = {'inputfile': 'tests/files/input_blackList.vcf', 'outputfile': 'tests/files/main_test.out',
-            'aftag': None, 'afthr': '0.2', 'bigfile': None, 'verbose': None}
+            'aftag': None, 'afthr': '0.2', 'bigfile': None, 'BEDfile': None, 'verbose': None}
     # Run and Tests
     with pytest.raises(SystemExit) as e:
         assert main_blackList(args)
@@ -106,7 +119,7 @@ def test_args_afthr_conflict():
 #     ''' '''
 #     # Variables
 #     args = {'inputfile': 'tests/files/input_blackList_novoAF_missing.vcf', 'outputfile': 'tests/files/main_test.out',
-#             'aftag': 'novoAF', 'afthr': '0.2', 'bigfile': None}
+#             'aftag': 'novoAF', 'afthr': '0.2', 'bigfile': None, 'BEDfile': None,}
 #     # Run and Tests
 #     with pytest.raises(SystemExit) as e:
 #         assert main_blackList(args)